Updated openstack/openstack Project: openstack/manila  cc6775d5945d0d76ceeec8b94dfc256fb9a4ea8b
Disable notifications

As part of the manila-telemetry integration work
we are now triggering notifications for several
events. We should provide a way to disable those.

In this patch-set we define a decorator that can be
used to change the behavior of those classes used
to emit notification.

Partially-Implements: bp ceilometer-integration

Change-Id: I806e0133e9fe3ad5cc35ad1e000e5ca61ff3b8ca