Updated openstack/openstack Project: openstack/diskimage-builder  d66dbc679c6dd44f3282aba36829432a19a50707
The correct option for label name in fat and vfat is '-n'

The code in mkfs correctly extends the command line with a '-n' for
vfat but does not currently do it for fat. This means that mkfs for
fat ends up with a '-L' which is what you'd do for everything like
ext[234].

The change just treats fat like vfat in the one place where this check
is required.

Change-Id: If65dfd949acdadff33a564640fb42ea73026a786
Closes-Bug: #1703063