Updated openstack/openstack Project: openstack/swift  5b10cf530bd1cb2624711f53e0ae45be30b70cb2
Add more structure to the deployment guide

Previously it was hard to navigate to a particular config section in
the deployment guide, and not possible to provide a link directly to
one section.

This patch makes each config section a heading so that it appears in
navigation tables and can be easily linked to. A list of config
sections is also added at the start of each server section.

Change-Id: Iecb0637fde521600a9163fa66b3dbdc176a71dff
Related-Bug: #1626290