Updated openstack/openstack Project: openstack/django_openstack_auth  acd690ae6c3c2e242c0c200ebac96c615a355d69
Allow for manual setting of default service region in config

In case DEFAULT_SERVICE_REGIONS setting in Horizon config is specified
(on a per-endpoint basis), use it instead of a value stored in
cookies. This value is still checked for sanity, i.e. it should be
present in Keystone service catalog.

Change-Id: Ia4787b56db7ce7787bd8aac21b5c0ec8a95a6f09
Related-Bug: #1506825
Closes-Bug: #1703390