Updated openstack/openstack Project: openstack/nova  e45accc5fd0c9ab995be3391f60b0ea57b08f8a4
Only setup iptables for metadata if using nova-net

As discussed in the bug report, we setup iptables rules for the metadata
service even if we're using neutron (which routes to metadata in a
different way). This is because of the split-brain behaviour of the
network driver interface versus the network API interface.

Instead, only setup iptables if we are _not_ using neutron.

Change-Id: I43df9200aba1018d2c7cd2f118864326af15fd42
Closes-Bug: #1687187