Updated openstack/openstack Project: openstack/python-novaclient  712b533dd0caf7c81f62a49cbce7f34a29cadd7f
Adjust test_resize_down_revert to account for counting quotas

Change I9269ffa2b80e48db96c622d0dc0817738854f602 in the Nova
server removes the need for reservations, which means that
once the server is resized in nova-compute and the vcpus and
memory_mb are changed on the instances record, the usage count
will also change, before the resize is confirmed. This means
we need to adjust the usage assertions in the resize down test.
Usage changes once the server is in VERIFY_RESIZE state, and
the usage will change again if the resize is reverted.

Note that if we were getting usage without the 'reserved'
query paramter to the 'limits' API, the behavior would have
been the same regardless of the change to Nova for counting
quotas.

Depends-On: I9269ffa2b80e48db96c622d0dc0817738854f602

Related to blueprint cells-count-resources-to-check-quota-in-api

Change-Id: I91ca0d204454987dcce1c68d9bf0f532b469df6d