Updated openstack/openstack Project: openstack-infra/project-config  a4c3e6e105e2dd312e93f378cf647efe16b3c73e
Let unbound find rc.local on Debian stretch in DIB

Building an image for stretch results in the unbound configuration
script throwing an error, that it cannot know where to find rc.local.
This happens, because the file is missing in the installed image.

The script now checks for /etc/debian_version instead of erroring out.
This file is available on any debian system. If there's no rc.local and
/etc/debian_version is available, rc.local can just be created with the
appropriate content, /etc/rc.local is executed on Debian.

Change-Id: If98f7ef1657b4f538a4f79ba116c8cff991aefc7