Updated openstack/openstack Project: openstack/barbican  d3b4dc31501da0990d99669d16366d69548b8d45
Add PBR wsgi_scripts entrypoint for barbican api

This commit makes the barbican wsgi script consistent with other
services by leveraging the pbr wsgi_scripts entrypoint to expose it.
While you could still leverage the bundled app.wsgi the entrypoint
barbican's behavior the same as other services. As part of this the
app.wsgi script is deprecated and will be removed in the future.

Change-Id: Id42f76dbfd59209232b37096a708ee18cbf96431