Updated openstack/openstack Project: openstack/cinder  45f0ab5e60668a35a95489a039a5b8265cab2f53
ZFSSA iSCSI delete volume with non-existent LUN

Under some circumstances, a volume can exist in cinder for which there
is no corresponding LUN on the ZFSSA. This fix allows deletion of the
volume in such cases, by catching the NOT_FOUND status from the ZFSSA
REST API and translating it to a VolumeNotFound exception, and then
considering that exception to be non-fatal, whilst passing up any
other exception that may be encountered. This way, some other failure
in using the REST API to obtain LUN information (e.g. transient
network failure) will not cause the cinder volume to get blindly
deleted, which would have left behind an orphaned LUN on the ZFSSA.

Change-Id: I332163fa35a2aa3f6f921b805fa97c803ec10724
Closes-Bug: #1537914