Updated openstack/openstack Project: openstack/heat  d85361bd9f44fda03cdb9e39f53c4b4180d3dd04
Add validation of block_device_mapping_v2

Heat assumes image/volume/snapshot in bdm_v2
as bootable sources, if user does not specify
the 'boot_index' for them, heat will give the
value '0' for all of them, but nova does not
allow this.
This patch adds validation to avoid the error.

Closes-Bug: #1658007
Change-Id: I6809c8cb839eba20ffe12f783854fbaa7b9267d6