Updated openstack/openstack Project: openstack/openstack-manuals  f81691fa010d85494f37e248dc11334b9dd25f58
replace pike index page with the main site version

Replace pike/index.html with content from the top level page in
preparation to make the top level page redirect to a page in a series
subdirectory.

Incorporate the content of various included templates directly into
the page to simplify understanding how the page is built and to change
some of the links to their new locations.

Change-Id: I93b5be93c5275bbed26199a8d44aa654d844877f
Signed-off-by: Doug Hellmann <doug@doughellmann.com>