Updated openstack/openstack Project: openstack/cinder  d79c107925c8bc04fd04f411de39c34b94db0742
VMAX driver - bootable volume errors

There are some issues in creating VMAX bootable volumes. This patch
rectifies these issues by correcting the unlink flow of the volume,
correcting the cleanup of the masking view, and adding a lock to the
volume to avoid concurrency issues. This patch also sets default
oversubscription in the event it is set to less than 1, and corrects
some minor punctuation.

Change-Id: If04542029db3a4835956d55644865e943a6f6fc0
Closes-Bug: #1696975