Updated openstack/openstack Project: openstack/keystone  68ee9ea5c44c9e899ee8d38c47d1dcddc99a9e87
fix assert_admin

The assert_admin method was broken by an earlier Pike commit [1]
which caused it to check for the "identity:admin_required" policy
rule (which is not defined) instead of the "admin_required" rule.
This changes check_policy to take the action instead of a function
name, allowing the caller to decide whether they want to prepend
"identity:" or not.

[1] 40c118d6ba94b15d6f66da618e1d368a0f876340

Change-Id: Ief05967361959a0bd194af3e77af198936e4a71f
Closes-Bug: #1703467