Updated openstack/openstack Project: openstack/tripleo-heat-templates  bb963cfe422d1ad4de2fa5b7cf24fa3907582ad9
Make Deploy/UpdateIdentifier definition semi-consistent

It seems UpdateIdentifier is an overloaded parameter - it is used
both to trigger package updates in the minor update case as well as
to trigger the upgrade steps during a major upgrade.  I'm not sure
it's appropriate to change either of the descriptions as a result,
so for the moment that is added to the exclusion list.

Change-Id: Ied36cf259f6a6e5c8cfa7a01722fb7fda6900976
Partial-Bug: 1700664