Updated openstack/openstack Project: openstack/nova  d2dbd4f5475fda95deba99809cfc085c7e44e1c5
[placement] cover deleting a custom resource class in use

If there is inventory or allocations of a custom resource class we
shouldn't be able to delete. This added gabbi file create a resource
provider and a resource class, sets some inventory, adds an allocation
and then tries to delete the resource class. When there is any inventory
the response should 409.

Change-Id: Iee3095cd23a4559d3c6a2f2202fe76e29bf7e9d9