Updated openstack/openstack Project: openstack/heat  6a42ff6c883f6800292fedafc2fb7995aae02fb3
Unit tests: make tests more realistic

In some tests the template we provide for the Stack does not match the
actual resources we are testing. This never happens in reality, and it
makes the tests brittle when other legitimate changes occur in the engine.
This makes the integration testing more realistic.

Change-Id: I777309f310e18f8a16b5ae675419d7eeeffa09ee