Updated openstack/openstack Project: openstack/tripleo-heat-templates  5cbcc8377c49e395dc1d02a976d9b4a94253f5ca
Remove non-containerized pacemaker resources on upgrade

Adds upgrade_tasks to remove the pacemaker resources using the
ansible-pacemaker module.

Resources are disabled and removed in step2 (called only on
bootstrap node) and then the cluster stop is moved to step3

The existing systemd/service call is kept but only to disable
services after they are disabled/deleted from the cluster.

Related-Bug: 1701485
Co-Authored-By: Damien Ciabrini <dciabrin@redhat.com>
Change-Id: Ia597d240ea5834c50a8f6c4fac0b6ed417b8535c