Updated openstack/openstack Project: openstack-infra/tripleo-ci  d3f7cf887a6439fc0604f37ddd30309f0c3326c4
Nic config for containerized undercloud

Adds a nic config template for the containerized undercloud. The
template creates a br-ctlplane bridge with no members, which matches the
same network configuration we use for all other multinode jobs.

Also sets the $LOCAL_IP (--local-ip) option to the default of
192.168.24.1 instead of reading the ip address off the interface with
the default route. Previously, endpoints were getting created with the
external IP address of the multinode instance, which resulted in
some issues (provisioning interface not assigned an IP in the same cidr
as the ctlplane subnet).

Change-Id: Ife080590e377d0e41881ba906be65908c62591b3