Updated openstack/openstack Project: openstack/nova  3998d6114ea1aa81066fa1330eb86a9d7700d527
placement: add retry tight loop claim_resources()

Adds a recursive call to the scheduler report client's
claim_resources() method that retries in the event (and only in the
event) of a concurrent update detected. If this error is returned, a
retry of the exact same allocation request is safe and perfectly
reasonable to attempt in normal operation.

Change-Id: Ib1abbb2deb1c0fd0a133e13c3cd2a967e3b3dc58
blueprint: placement-allocation-requests