Updated openstack/openstack Project: openstack/tripleo-heat-templates  13c936c067807de5e6a330c4944408313b093bb2
Mount /var/lib/neutron in neutron agents for metadata proxy

The metadata agent creates domain socket /var/lib/neutron/metadata_proxy
that is used for communication with haproxy in the L3 and DHCP agents.
This patch adds creation of /var/lib/neutron if it doesn't exist and
mounts it into the L3, DHCP and metadata agent containers.

Change-Id: Id8b8487b5a6a288e5ef1ca1c7d5b47a59cc8dea2
Closes-Bug: #1705289