Updated openstack/openstack Project: openstack-infra/project-config  0a3bdc462fcac5e4136d6e8c60048c95217998bf
Add a simple version of Mistral HA gate

The idea of the new gate is to run tests on multi-node Mistral w/o
Devstack and Tempest at all. For now the gate configuration is
very primitive and just calls two scripts from the mistral repo
so that we could debug more easily by sending patchsets to Mistral
with changes made against these scripts. Once they work it'll be
better to create a builder within the Jenkins Mistral configuration.

Change-Id: I898e125d37a31a00e972be5962a49af2a6076acd