Updated openstack/openstack Project: openstack/glance  309cf3a6ed3948aad39b9a9b884fb3fee969d29b
Fix filter doesn't support non-ascii characters

When non-ascii characters is passed as filter parameter to "image-list"
and "md-namespace-list" APIs then it returns HTTP 500 if glance service
is running using Python2. This is because query parameters passed to
the urlparse.urlencode method raises UnicodeEncodeError for non-ascii
characters.

Overridden params property of webob.request.BaseRequest in wsgi layer
Request class to avoid encoding of request params at several places in
case of Py2. Added a new class attribute 'encoded_params' which is
initialized to encoded params dict when params property is called for
the first time and in the subsequent calls this attribute is returned
without requiring you to worry whether the params value are encoded
or not.

Closes-Bug: #1521193
Change-Id: If4497d42820aa4b8070666d2a1a4413411557f12