Updated openstack/openstack Project: openstack/cinder  edf773125fe090d4b3a2ce8853bac9667279c7f4
Mark ZTE driver as unsupported

The ZTE CI has not reported success past our CI required period.

Checking name: ZTE cinder2 CI
  first seen: 2017-01-09 14:41:08 (154 days, 0:31:49 old)
  last seen: 2017-03-01 05:53:37 (103 days, 9:19:20 old)
  last success: 2017-02-13 13:00:31 (119 days, 2:12:26 old)
  Job zte-cinder-tempest 9% success out of 182 comments S=17, F=165
      last success: 2017-02-13 13:00:31 (119 days, 2:12:26 old)
Overall success rate: 9% of 250 comments

Per Cinder's non-compliance policy, this patch marks the driver
as unsupported and deprecated and it will be removed in the next
cycle if the issue is not corrected.

Change-Id: Iaa82848bff503bde00cb6abaea94445529ecc6ee