Updated openstack/openstack Project: openstack/heat  4d6dfbb8adfc1ebfc96c5db0ea897a771cc9d84d
Don't call Stack.has_cache_data() in Functions

This check was added in 063803067178bc26f9e3c1f512ffee402bc59d50 because
originally the Resource objects other than the one we were checking were
not initialised with any data during a convergence resource check.
Therefore they would be in the INIT_COMPLETE state, but would return valid
attribute values from the cache data.

It was made redundant by 1956ddd2a61a69ab28cdf0645c4b1973be9adbce, which
loaded the resource state in the Resource objects. The state will never be
INIT_COMPLETE.

This code requires the Function's 'stack' member to be a Stack object, so
we have to get rid of it in order to be able to start using a
StackDefinition object in its place.

Change-Id: I1a97401bf126a3ec10cc420bbb0786b785035d48