Updated openstack/openstack Project: openstack/nova  15500b89be5772be8925bcab9706687682487c93
Set IronicNodeState.uuid in _update_from_compute_node

The IronicNodeState extends the HostState class but failed
to set the uuid attribute, which is needed later when we
make allocation request claims to Placement in the FilterScheduler.

It seems like IronicNodeState._update_from_compute_node should
call super(IronicNodeState, self)._update_from_compute_node(compute)
so we don't need to worry about all of this duplication, but that
refactor is left for another change.

Part of blueprint placement-claims

Change-Id: I6210a7f0d5a0c71eb7cd9b9bd521b482d3aa9c66