Updated openstack/openstack Project: openstack/tempest  80046606ea05e33a4e16b4edaafd003e0813a956
Ensure test for Neutron GET / uses right URL

Previously, the logic in the API test for Neutron GET / did not always
determine the form of the unversioned endpoint correctly. For example,
when running Neutron on a shared port instead of on its own service
port, e.g. http://host/network/v2.0 as the endpoint, the test would
incorrectly run against http://host/ instead of http://host/network.

Now, ensure that the correct form of the unversioned endpoint is always
used in the test.

Change-Id: I9dcd30fe6669bdf570149ed317b16005c21f6291
Closes-Bug: #1703959