Updated openstack/openstack Project: openstack/openstackdocstheme  979785c760d6d4ffe46fa09037844657a32eb727
Fix a logic issue in the JS

When I updated the icon selector as part of
https://review.openstack.org/#/c/485290, I made an error in the logic
meaning that the icons appear next to only the first of each item.
This was part of an initial proposed solution, but isn't actually needed,
and causes multiple usages of versionadded etc to not have icons.

Change-Id: If1c1011d8a7ed1e78f2675eca7daf4e497a8bf9e