Updated openstack/openstack Project: openstack/tripleo-heat-templates  95a6da61129ec6a88836f99c8f2d0260a53bba96
Stop also openstack-swift-object-expirer when upgrading swift services

openstack-swift-object-expirer is not stopped when
running the upgrade tasks so forth when changing to
containers the service is still running after upgrading
to docker.

This service is added by default here: https://review.openstack.org/#/c/404149
But it wasnt stopped when running the upgrade tasks.

Related also to this RHBZ#1470005

Change-Id: I8d5f195095d702057c3b2741127e7338d7451aad
Closes-Bug: 1699444